requirements.txt
aniso8601==8.0.0
api==0.0.7
argcomplete==1.10.0
attrs==20.1.0
beautifulsoup4==4.8.0
blis==0.4.1
catalogue==1.0.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cymem==2.0.3
docx2txt==0.8
EbookLib==0.17.1
extract-msg==0.23.1
Flask==1.1.2
Flask-Cors==3.0.9
flask-restplus==0.13.0
idna==2.10
IMAPClient==2.1.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
lxml==4.5.2
MarkupSafe==1.1.1
murmurhash==1.0.2
nose==1.3.7
numpy==1.19.1
olefile==0.46
pdfminer.six==20181108
Pillow==7.2.0
plac==1.1.3
preshed==3.0.2
pycryptodome==3.9.8
pyrsistent==0.16.0
python-pptx==0.6.18
pytz==2020.1
requests==2.24.0
six==1.12.0
sortedcontainers==2.2.2
soupsieve==2.0.1
spacy==2.3.2
SpeechRecognition==3.8.1
srsly==1.0.2
textract==1.6.3
thinc==7.4.1
tqdm==4.48.2
tzlocal==1.5.1
urllib3==1.25.10
wasabi==0.8.0
Werkzeug==0.16.0
xlrd==1.2.0
XlsxWriter==1.3.3
